+++ tests/query_element/elements.py
@@ -21,7 +21,7 @@ class ProductSearch(BaseProjection):
 
 class Money(BaseValueObject):
     amount = Float(required=True)
-    currency = String(required=True, max_length=3, default="USD")
+    currency = String(max_length=3, default="USD")
 
 
 class GetOrdersByCustomer(BaseQuery):